refactor(#4754): extract manual steer mailbox claim
